requirements.txt
typer[all]==0.12.5
rich==13.7.1
groq==0.13.0
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.6
setuptools==69.2.0